+++ lib/heroicons/mini.ex
@@ -1,9 +1,0 @@
-defmodule Heroicons.Mini do
-  @moduledoc """
-  Solid style icons drawn with fills, packaged as Phoenix Components.
-  For smaller elements like buttons, form elements, and to support text,
-  designed to be rendered at 20x20.
-  """
-
-  use Heroicons.Generator, icon_dir: "icons/mini/"
-end