+++ test/heroicons_test.exs
@@ -3,67 +3,50 @@ defmodule HeroiconsTest do
 
   import Phoenix.LiveViewTest
 
-  test "generated function" do
-    result = Phoenix.HTML.safe_to_string(Heroicons.Outline.academic_cap())
-    assert result =~ ~s(<svg )
-    assert result =~ ~s(aria-hidden="true")
-    assert result =~ ~s(fill="none")
-    assert result =~ ~s(stroke="currentColor")
-    assert result =~ ~s(stroke-width="1.5")
-    assert result =~ ~s(viewBox="0 0 24 24")
-
-    assert Heroicons.Outline.academic_cap(class: "w-6 h-6 text-gray-500")
-           |> Phoenix.HTML.safe_to_string() =~
-             ~s(class="w-6 h-6 text-gray-500")
-
-    assert Heroicons.Outline.academic_cap(class: "<> \" ")
-           |> Phoenix.HTML.safe_to_string() =~
-             ~s(class="&lt;&gt; &quot; ")
-
-    assert Heroicons.Outline.academic_cap(foo: "bar")
-           |> Phoenix.HTML.safe_to_string() =~
-             ~s(foo="bar")
-
-    assert Heroicons.Outline.academic_cap(multiword_key: "foo")
-           |> Phoenix.HTML.safe_to_string() =~
-             ~s(multiword-key="foo")
-
-    assert Heroicons.Outline.academic_cap(viewBox: "0 0 12 12")
-           |> Phoenix.HTML.safe_to_string() =~
-             ~s(viewBox="0 0 12 12")
+  setup do
+    Mix.Task.run("heroicons.build")
+    {:ok, assigns: %{__changed__: nil}}
   end
 
-  test "generated components" do
-    outline = render_component(&Heroicons.Outline.academic_cap/1, assigns())
-    assert outline =~ ~s(<svg )
-    assert outline =~ ~s(aria-hidden="true")
-    assert outline =~ ~s(fill="none")
-    assert outline =~ ~s(stroke="currentColor")
-    assert outline =~ ~s(stroke-width="1.5")
-    assert outline =~ ~s(viewBox="0 0 24 24")
+  test "outline default" do
+    assert render_component(&Heroicons.cake/1, %{}) ==
+             "<svg xmlns=\"http://www.w3.org/2000/svg\" aria-hidden=\"true\" fill=\"none\" viewBox=\"0 0 24 24\" stroke-width=\"1.5\" stroke=\"currentColor\">\n  <path stroke-linecap=\"round\" stroke-linejoin=\"round\" d=\"M12 8.25v-1.5m0 1.5c-1.355 0-2.697.056-4.024.166C6.845 8.51 6 9.473 6 10.608v2.513m6-4.87c1.355 0 2.697.055 4.024.165C17.155 8.51 18 9.473 18 10.608v2.513m-3-4.87v-1.5m-6 1.5v-1.5m12 9.75l-1.5.75a3.354 3.354 0 01-3 0 3.354 3.354 0 00-3 0 3.354 3.354 0 01-3 0 3.354 3.354 0 00-3 0 3.354 3.354 0 01-3 0L3 16.5m15-3.38a48.474 48.474 0 00-6-.37c-2.032 0-4.034.125-6 .37m12 0c.39.049.777.102 1.163.16 1.07.16 1.837 1.094 1.837 2.175v5.17c0 .62-.504 1.124-1.125 1.124H4.125A1.125 1.125 0 013 20.625v-5.17c0-1.08.768-2.014 1.837-2.174A47.78 47.78 0 016 13.12M12.265 3.11a.375.375 0 11-.53 0L12 2.845l.265.265zm-3 0a.375.375 0 11-.53 0L9 2.845l.265.265zm6 0a.375.375 0 11-.53 0L15 2.845l.265.265z\"/>\n</svg>"
 
-    solid = render_component(&Heroicons.Solid.academic_cap/1, assigns())
-    assert solid =~ ~s(<svg )
-    assert solid =~ ~s(aria-hidden="true")
-    assert solid =~ ~s(fill="currentColor")
-    assert solid =~ ~s(viewBox="0 0 24 24")
+    class = "w-2"
 
-    assert render_component(
-             &Heroicons.Outline.academic_cap/1,
-             assigns(class: "w-6 h-6 text-gray-500")
-           ) =~
-             ~s(class="w-6 h-6 text-gray-500")
+    assert render_component(&Heroicons.cake/1, %{class: class}) ==
+             "<svg xmlns=\"http://www.w3.org/2000/svg\" aria-hidden=\"true\" fill=\"none\" viewBox=\"0 0 24 24\" stroke-width=\"1.5\" stroke=\"currentColor\" class=\"#{class}\">\n  <path stroke-linecap=\"round\" stroke-linejoin=\"round\" d=\"M12 8.25v-1.5m0 1.5c-1.355 0-2.697.056-4.024.166C6.845 8.51 6 9.473 6 10.608v2.513m6-4.87c1.355 0 2.697.055 4.024.165C17.155 8.51 18 9.473 18 10.608v2.513m-3-4.87v-1.5m-6 1.5v-1.5m12 9.75l-1.5.75a3.354 3.354 0 01-3 0 3.354 3.354 0 00-3 0 3.354 3.354 0 01-3 0 3.354 3.354 0 00-3 0 3.354 3.354 0 01-3 0L3 16.5m15-3.38a48.474 48.474 0 00-6-.37c-2.032 0-4.034.125-6 .37m12 0c.39.049.777.102 1.163.16 1.07.16 1.837 1.094 1.837 2.175v5.17c0 .62-.504 1.124-1.125 1.124H4.125A1.125 1.125 0 013 20.625v-5.17c0-1.08.768-2.014 1.837-2.174A47.78 47.78 0 016 13.12M12.265 3.11a.375.375 0 11-.53 0L12 2.845l.265.265zm-3 0a.375.375 0 11-.53 0L9 2.845l.265.265zm6 0a.375.375 0 11-.53 0L15 2.845l.265.265z\"/>\n</svg>"
+  end
 
-    mini = render_component(&Heroicons.Mini.academic_cap/1, assigns())
-    assert mini =~ ~s(<svg )
-    assert mini =~ ~s(aria-hidden="true")
-    assert mini =~ ~s(fill="currentColor")
-    assert mini =~ ~s(viewBox="0 0 20 20")
+  test "solid" do
+    assert render_component(&Heroicons.cake/1, %{solid: true}) ==
+             "<svg xmlns=\"http://www.w3.org/2000/svg\" aria-hidden=\"true\" viewBox=\"0 0 24 24\" fill=\"currentColor\">\n  <path d=\"M15 1.784l-.796.796a1.125 1.125 0 101.591 0L15 1.784zM12 1.784l-.796.796a1.125 1.125 0 101.591 0L12 1.784zM9 1.784l-.796.796a1.125 1.125 0 101.591 0L9 1.784zM9.75 7.547c.498-.02.998-.035 1.5-.042V6.75a.75.75 0 011.5 0v.755c.502.007 1.002.021 1.5.042V6.75a.75.75 0 011.5 0v.88l.307.022c1.55.117 2.693 1.427 2.693 2.946v1.018a62.182 62.182 0 00-13.5 0v-1.018c0-1.519 1.143-2.829 2.693-2.946l.307-.022v-.88a.75.75 0 011.5 0v.797zM12 12.75c-2.472 0-4.9.184-7.274.54-1.454.217-2.476 1.482-2.476 2.916v.384a4.104 4.104 0 012.585.364 2.605 2.605 0 002.33 0 4.104 4.104 0 013.67 0 2.605 2.605 0 002.33 0 4.104 4.104 0 013.67 0 2.605 2.605 0 002.33 0 4.104 4.104 0 012.585-.364v-.384c0-1.434-1.022-2.7-2.476-2.917A49.138 49.138 0 0012 12.75zM21.75 18.131a2.604 2.604 0 00-1.915.165 4.104 4.104 0 01-3.67 0 2.604 2.604 0 00-2.33 0 4.104 4.104 0 01-3.67 0 2.604 2.604 0 00-2.33 0 4.104 4.104 0 01-3.67 0 2.604 2.604 0 00-1.915-.165v2.494c0 1.036.84 1.875 1.875 1.875h15.75c1.035 0 1.875-.84 1.875-1.875v-2.494z\"/>\n</svg>"
+
+    class = "w-2"
+
+    assert render_component(&Heroicons.cake/1, %{solid: true, class: class}) ==
+             "<svg xmlns=\"http://www.w3.org/2000/svg\" aria-hidden=\"true\" viewBox=\"0 0 24 24\" fill=\"currentColor\" class=\"w-2\">\n  <path d=\"M15 1.784l-.796.796a1.125 1.125 0 101.591 0L15 1.784zM12 1.784l-.796.796a1.125 1.125 0 101.591 0L12 1.784zM9 1.784l-.796.796a1.125 1.125 0 101.591 0L9 1.784zM9.75 7.547c.498-.02.998-.035 1.5-.042V6.75a.75.75 0 011.5 0v.755c.502.007 1.002.021 1.5.042V6.75a.75.75 0 011.5 0v.88l.307.022c1.55.117 2.693 1.427 2.693 2.946v1.018a62.182 62.182 0 00-13.5 0v-1.018c0-1.519 1.143-2.829 2.693-2.946l.307-.022v-.88a.75.75 0 011.5 0v.797zM12 12.75c-2.472 0-4.9.184-7.274.54-1.454.217-2.476 1.482-2.476 2.916v.384a4.104 4.104 0 012.585.364 2.605 2.605 0 002.33 0 4.104 4.104 0 013.67 0 2.605 2.605 0 002.33 0 4.104 4.104 0 013.67 0 2.605 2.605 0 002.33 0 4.104 4.104 0 012.585-.364v-.384c0-1.434-1.022-2.7-2.476-2.917A49.138 49.138 0 0012 12.75zM21.75 18.131a2.604 2.604 0 00-1.915.165 4.104 4.104 0 01-3.67 0 2.604 2.604 0 00-2.33 0 4.104 4.104 0 01-3.67 0 2.604 2.604 0 00-2.33 0 4.104 4.104 0 01-3.67 0 2.604 2.604 0 00-1.915-.165v2.494c0 1.036.84 1.875 1.875 1.875h15.75c1.035 0 1.875-.84 1.875-1.875v-2.494z\"/>\n</svg>"
+  end
+
+  test "mini" do
+    assert render_component(&Heroicons.cake/1, %{mini: true}) ==
+             "<svg xmlns=\"http://www.w3.org/2000/svg\" aria-hidden=\"true\" viewBox=\"0 0 20 20\" fill=\"currentColor\">\n  <path d=\"M6.75.98l-.884.883a1.25 1.25 0 101.768 0L6.75.98zM13.25.98l-.884.883a1.25 1.25 0 101.768 0L13.25.98zM10 .98l.884.883a1.25 1.25 0 11-1.768 0L10 .98zM7.5 5.75a.75.75 0 00-1.5 0v.464c-1.179.305-2 1.39-2 2.622v.094c.1-.02.202-.038.306-.051A42.869 42.869 0 0110 8.5c1.93 0 3.83.129 5.694.379.104.013.206.03.306.051v-.094c0-1.232-.821-2.317-2-2.622V5.75a.75.75 0 00-1.5 0v.318a45.645 45.645 0 00-1.75-.062V5.75a.75.75 0 00-1.5 0v.256c-.586.01-1.17.03-1.75.062V5.75zM4.505 10.365A41.377 41.377 0 0110 10c1.863 0 3.697.124 5.495.365C16.967 10.562 18 11.838 18 13.28v.693a3.72 3.72 0 01-1.665-.393 5.222 5.222 0 00-4.67 0 3.722 3.722 0 01-3.33 0 5.222 5.222 0 00-4.67 0A3.72 3.72 0 012 13.972v-.693c0-1.441 1.033-2.716 2.505-2.914zM15.665 14.921a5.22 5.22 0 002.335.551V16.5a1.5 1.5 0 01-1.5 1.5h-13A1.5 1.5 0 012 16.5v-1.028c.8 0 1.6-.183 2.335-.551a3.722 3.722 0 013.33 0c1.47.735 3.2.735 4.67 0a3.722 3.722 0 013.33 0z\"/>\n</svg>"
+
+    class = "w-2"
+
+    assert render_component(&Heroicons.cake/1, %{mini: true, class: "w-2"}) ==
+             "<svg xmlns=\"http://www.w3.org/2000/svg\" aria-hidden=\"true\" viewBox=\"0 0 20 20\" fill=\"currentColor\" class=\"#{class}\">\n  <path d=\"M6.75.98l-.884.883a1.25 1.25 0 101.768 0L6.75.98zM13.25.98l-.884.883a1.25 1.25 0 101.768 0L13.25.98zM10 .98l.884.883a1.25 1.25 0 11-1.768 0L10 .98zM7.5 5.75a.75.75 0 00-1.5 0v.464c-1.179.305-2 1.39-2 2.622v.094c.1-.02.202-.038.306-.051A42.869 42.869 0 0110 8.5c1.93 0 3.83.129 5.694.379.104.013.206.03.306.051v-.094c0-1.232-.821-2.317-2-2.622V5.75a.75.75 0 00-1.5 0v.318a45.645 45.645 0 00-1.75-.062V5.75a.75.75 0 00-1.5 0v.256c-.586.01-1.17.03-1.75.062V5.75zM4.505 10.365A41.377 41.377 0 0110 10c1.863 0 3.697.124 5.495.365C16.967 10.562 18 11.838 18 13.28v.693a3.72 3.72 0 01-1.665-.393 5.222 5.222 0 00-4.67 0 3.722 3.722 0 01-3.33 0 5.222 5.222 0 00-4.67 0A3.72 3.72 0 012 13.972v-.693c0-1.441 1.033-2.716 2.505-2.914zM15.665 14.921a5.22 5.22 0 002.335.551V16.5a1.5 1.5 0 01-1.5 1.5h-13A1.5 1.5 0 012 16.5v-1.028c.8 0 1.6-.183 2.335-.551a3.722 3.722 0 013.33 0c1.47.735 3.2.735 4.67 0a3.722 3.722 0 013.33 0z\"/>\n</svg>"
+  end
+
+  test "raises on mix styles" do
+    assert_raise ArgumentError, "expected either mini or solid, but got both.", fn ->
+      render_component(&Heroicons.cake/1, %{mini: true, solid: true})
+    end
   end
 
   test "generated docs" do
     {:docs_v1, _annotation, _beam_language, _format, _module_doc, _metadata, docs} =
-      Code.fetch_docs(Heroicons.Outline)
+      Code.fetch_docs(Heroicons)
 
     doc =
       Enum.find_value(docs, fn
@@ -72,24 +55,27 @@ defmodule HeroiconsTest do
       end)
 
     assert doc["en"] == """
-           ![](assets/icons/outline/academic-cap.svg) {: width=24px}
+           Renders the `academic_cap` icon.
+
+           By default, the outlined (24x24) component is used, but the `solid` or `mini`
+           attributes can be provided for alternative styles.
+
+           You may also pass arbitrary HTML attributes to be applied to the svg tag.
 
            ## Examples
 
-           Use as a `Phoenix.Component`
+           ```heex
+           <Heroicons.academic_cap />
+           <Heroicons.academic_cap class="w-4 h-4" />
+           <Heroicons.academic_cap solid />
+           <Heroicons.academic_cap mini />
+           ```
 
-               <.academic_cap />
-               <.academic_cap class="w-6 h-6 text-gray-500" />
+           ## Attributes
 
-           or as a function
-
-               <%= academic_cap() %>
-               <%= academic_cap(class: "w-6 h-6 text-gray-500") %>
+           * `rest` (`:global`) - the arbitrary HTML attributes for the svg container.
+           * `solid` (`:boolean`) - Defaults to `false`.
+           * `mini` (`:boolean`) - Defaults to `false`.
            """
-  end
-
-  defp assigns(assgns \\ []) do
-    Map.new(assgns)
-    |> Map.put_new(:__changed__, %{})
   end
 end